requirements.txt
Flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
google-generativeai==0.3.2
torch==2.1.0
transformers==4.35.0
gunicorn==21.2.0
python-dotenv==1.0.0
